* 音视频伪直播: (1166 commits)
  伪直播：MP3播放时暂停再开始，转盘出来的问题
  版本号 3.5.6
  上台四人和五人的变化
  伪直播老师默认不上台
  音视频伪直播判断枚举
  直播中学生视频点击除的控制view独立出来
  no message
  roomModel中获取直播类型
  上台人数4变5
  no message
  no message
  去掉fakeView
  伪直播
  多摄像头streamId获取问题
  隐私政策下移
  学生左右布局没有复制窗口尺寸，导致视频窗口捏合缩放崩溃
  播放回放时默认使用外放声音，解决部分用户播放回放时没有声音问题
  BMKit
  YSLiveSDK 3.5.5.1
  BMKit
  ...

# Conflicts:
#	BMKit/BMKit/BMKit.h
#	BMKit/BMKit/Category/UIKit/UIImage+BMCategory.h
#	BMKit/BMKit/ThirdPary/TZImagePickerController/BMTZImagePickerController.h
#	YSLib/CHWhiteBoard/CHWhiteBoardResources.bundle/react_mobile_new_publishdir/main.6309daea.js
#	YSLib/CHWhiteBoard/CHWhiteBoardResources.bundle/react_mobile_new_publishdir/main.6309daea.js.map
#	YSLib/YSRoomSDK.framework/Info.plist
#	YSLib/YSRoomSDK.framework/YSRoomSDK
#	YSLib/YSWhiteBoard/YSWhiteBoard.framework/Headers/YSWhiteBoardManager.h
#	YSLib/YSWhiteBoard/YSWhiteBoard.framework/Info.plist
#	YSLib/YSWhiteBoard/YSWhiteBoard.framework/YSWhiteBoard
#	YSLib/YSWhiteBoard/YSWhiteBoardResources.bundle/react_mobile_new_publishdir/index.html
#	YSLive.xcodeproj/project.pbxproj
#	YSLive/Info.plist
#	YSLive/Login/YSLoginVC.m
#	YSLive/YSTeacherRole/YSTeacherRoleMainVC.m
#	YSLiveManager/YSLiveMacros.h
#	YSLiveManager/YSLiveManager.h
#	YSLiveManager/YSLiveManager.m
#	YSLiveSDK/YSLiveSDK/YSSDKManager.h
#	YSLiveSDK/YSLiveSDK/YSSDKManager.m
#	YSLiveSDK/YS_iOSSDK版本更新说明.docx
#	YSLiveSample/YSLiveSample/Login/YSLoginVC.m
#	YSOnlineSchool/Class/YSClassDayList.m

+++ YSLive/YSTools/OCBarrage/YSBarrageCell.h
@@ -19,9 +19,9 @@ NS_ASSUME_NONNULL_BEGIN
 @property (nonatomic, assign) NSTimeInterval idleTime;//开始闲置的时间, 闲置超过5秒的, 自动回收内存
 @property (nonatomic, strong, nullable) YSBarrageDescriptor *barrageDescriptor;
 @property (nonatomic, strong, readonly, nullable) CAAnimation *barrageAnimation;
-@property (nonatomic, assign) int trackIndex;
+@property (nonatomic, assign) NSInteger trackIndex;
 
-- (void)addBarrageAnimationWithDelegate:(id<CAAnimationDelegate>)animationDelegate;
+- (void)addBarrageAnimationWithDelegate:(id <CAAnimationDelegate>)animationDelegate;
 - (void)prepareForReuse;
 - (void)clearContents;
